feat(MaxOneMarker) reintroduced MaxOneMarker trait
`def_headers` will automatically implement it for all
headers which have `MAX_ONE=true` (and it should be
only implemented for them).

`get_single` now requires that the `HeaderKind` implements
the `MaxOneMarker` trait making it impossible to use it
in a bugprone way (i.e. using `get_single` with any non
`MAX_ONE=true` trait is extreamly likely a bug)